Graph-Agentic-RAG/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
networkx>=3.0
langchain-core>=0.2.0
langgraph>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
websockets>=12.0
requests>=2.31.0
httpx>=0.25.0
